+++ src/widgets/TbButtonGroupColumn.php
@@ -102,7 +102,7 @@ class TbButtonGroupColumn extends TbButtonColumn
 		}
 
 		if (isset($button['icon'])) {
-			if (strpos($button['icon'], 'icon') === false) {
+			if (strpos($button['icon'], 'icon') === false && strpos($button['icon'], 'fa') === false) {
 				$button['icon'] = 'icon-' . implode(' icon-', explode(' ', $button['icon']));
 			}