+++ src/widgets/TbDataColumn.php
@@ -36,11 +36,13 @@ class TbDataColumn extends CDataColumn
 			$sort = $this->grid->dataProvider->getSort();
 			$label = isset($this->header) ? $this->header : $sort->resolveLabel($this->name);
 
+            $booster = Bootstrap::getBooster();
+
 			if ($sort->resolveAttribute($this->name) !== false){
                             if($sort->getDirection($this->name) === CSort::SORT_ASC){
-                                $label .= ' <span class="icon-sort-down"></span>';
+                                $label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-asc' : 'icon-chevron-down').'"></span>';
                             } elseif($sort->getDirection($this->name) === CSort::SORT_DESC){
-                                $label .= ' <span class="icon-sort-up"></span>';
+                                $label .= ' <span class="'.($booster->fontAwesomeCss ? 'fa fa-sort-desc' : 'icon-chevron-up').'"></span>';
                             } else {
                                 $label .= ' ';
                             }