+++ src/widgets/TbDropdown.php
@@ -48,9 +48,8 @@ class TbDropdown extends TbBaseMenu
 	protected function renderMenuItem($item)
 	{
 		if (isset($item['icon'])) {
-			if (strpos($item['icon'], 'icon') === false) {
-				$pieces = explode(' ', $item['icon']);
-				$item['icon'] = 'icon-' . implode(' icon-', $pieces);
+			if (strpos($item['icon'], 'icon') === false && strpos($item['icon'], 'fa') === false) {
+				$item['icon'] = 'icon-' . implode(' icon-', explode(' ', $item['icon']));
 			}
 
 			$item['label'] = '<i class="' . $item['icon'] . '"></i> ' . $item['label'];