+++ src/widgets/TbGroupButtonColumn.php
@@ -53,7 +53,7 @@ class TbGroupButtonColumn extends CButtonColumn
 		}
 
 		if (isset($button['icon'])) {
-			if (strpos($button['icon'], 'icon') === false) {
+			if (strpos($button['icon'], 'icon') === false && strpos($button['icon'], 'fa') === false) {
 				$button['icon'] = 'icon-' . implode(' icon-', explode(' ', $button['icon']));
 			}